pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = ""
authors = ["yootako <yootako03@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
yt-dlp = "^2024.11.4"
python-dotenv = "^1.0.1"
discord-py = {extras = ["voice"], version = "^2.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
